+++ weldx/core.py
@@ -23,7 +23,9 @@ class MathematicalExpression:
     """Mathematical expression using sympy syntax."""
 
     def __init__(
-        self, expression: Union[sympy.Expr, str], parameters: Union[Dict, None] = None
+        self,
+        expression: Union[sympy.Expr, str],
+        parameters: Dict[str, Union[pint.Quantity, str]] = None,
     ):
         """Construct a MathematicalExpression.
 
@@ -50,6 +52,7 @@ class MathematicalExpression:
                 raise ValueError(
                     f'"parameters" must be dictionary, got {type(parameters)}'
                 )
+            parameters = {k: Q_(v) for k, v in parameters.items()}
             variable_names = self.get_variable_names()
             for key in parameters:
                 if key not in variable_names:
@@ -354,6 +357,7 @@ class TimeSeries:
             interpolation = "step"
 
         # expand dim for scalar input
+        data = Q_(data)
         if not np.iterable(data):
             data = np.expand_dims(data, 0)