+++ CHANGELOG.md
@@ -18,9 +18,7 @@
 - Add `is_discrete` and `is_expression` properties to `TimeSeries` [[#366]](https://github.com/BAMWelDX/weldx/pull/366)
 - Add `MeasurementChain.output_signal` property that returns the output signal of the `MeasurementChain`
   [[#394]](https://github.com/BAMWelDX/weldx/pull/394)
-
-
-
+  
 ### changes
 
 - `WXRotation.from_euler()` now accepts a `pint.Quantity` as input. [[#318]](https://github.com/BAMWelDX/weldx/pull/318)
@@ -40,6 +38,8 @@
     - Added additional functionality to `Signal`, `SignalTransformation` and `GenericEquipment`
     - Removed `Data` class
     - Updated asdf schemas of all modified classes and the ones that contained references to those classes
+- allow input of string quantities in `MathematicalExpression` parameters and a few other
+  places [[#402]](https://github.com/BAMWelDX/weldx/pull/402)
 - `LocalCoordinateSystem.__init__` now accepts a `TimeSeries` as input. All methods of the `CoordinateSystemManager`
   also support this new behavior [[#366]](https://github.com/BAMWelDX/weldx/pull/366)